extract classes, restructure files
